WIP: build tool as a binaryTarget

+++ Sources/libmcpp/directive.c
@@ -112,7 +112,7 @@ void    directive( void)
     FILEINFO *  file;
     int     token_type;
     int     hash;
-    int     c;
+    long     c;
     char *  tp;
 
     in_directive = TRUE;
