+++ Sources/libmcpp/support.c
@@ -1943,7 +1943,7 @@ static char *   get_line(
 #endif
     static int  cr_converted;
     int     converted = FALSE;
-    int     len;                            /* Line length - alpha  */
+    size_t  len;                            /* Line length - alpha  */
     char *  ptr;
     int     cat_line = 0;           /* Number of catenated lines    */
 
@@ -1997,7 +1997,7 @@ static char *   get_line(
             /* Translation phase 2  */
             len -= 2;
             if (len >= 0) {
-                if ((*(ptr + len) == '\\') && ! last_is_mbchar( ptr, len)) {
+                if ((*(ptr + len) == '\\') && ! last_is_mbchar( ptr, (int)len)) {
                             /* <backslash><newline> (not MBCHAR)    */
                     ptr = infile->bptr += len;  /* Splice the lines */
                     wrong_line = TRUE;
@@ -2090,7 +2090,7 @@ int     cnv_digraph(
  */
 {
     int     count = 0;
-    int     i;
+    size_t  i;
     int     c1, c2;
 
     while ((i = strcspn( in, "%:<")), (c1 = *(in + i)) != '\0') {
